[refactor] Move Context to taichi/program
