+++ taichi/program/context.h
@@ -4,8 +4,8 @@
 #include "taichi/inc/constants.h"
 
 #if defined(TI_RUNTIME_HOST)
-namespace taichi::lang {
-using namespace taichi;
+namespace taichi {
+namespace lang {
 #endif
 
 struct LLVMRuntime;
@@ -36,21 +36,7 @@ struct Context {
 #endif
 };
 
-struct MemRequest {
-  std::size_t size;
-  std::size_t alignment;
-  uint8 *ptr;
-  std::size_t __padding;
-};
-
-static_assert((sizeof(MemRequest) & (sizeof(MemRequest) - 1)) == 0);
-
-struct MemRequestQueue {
-  MemRequest requests[taichi_max_num_mem_requests];
-  int tail;
-  int processed;
-};
-
 #if defined(TI_RUNTIME_HOST)
-}
+}  // namespace lang
+}  // namespace taichi
 #endif
